--- __init__.py
@@ -1,2 +1,0 @@
-# encoding: utf-8
-__author__ = 'liuzhijun'
