--- test.py
@@ -1,9 +1,0 @@
-import unittest
-
-from .zhihu import Zhihu
-
-class ZhihuTestCase(unittest.TestCase):
-    def test_user(self):
-        pass
-
-    pass
