+++ zhihu/__init__.py
@@ -1,0 +1,15 @@
+# encoding: utf-8
+__author__ = 'liuzhijun'
+
+# encoding: utf-8
+# !/usr/bin/env python
+
+"""
+知乎API
+"""
+
+from .models import answer
+from .models import zhihu
+
+Answer = answer.Answer
+Zhihu = zhihu.Zhihu
